# storage: filesystem/dotgit, Improve load packed-refs