In this PR I tweaked the parsing strategy of packed-refs to not continue if it is a single reference if it has already been found, which is very useful for repositories with thousands of references.
This reduces one iteration of references for loading all references, and also reduces memory allocations for a large number of references.