@fcharlie 2023-04-25
> Please reword the commit to align with [contributing guidelines](https://github.com/go-git/go-git/blob/master/CONTRIBUTING.md). Or refer to the new PR's title.
> 
> Please add a test to safeguard the changes against regression. It would be also nice if you could paste the benchmark results in the PR to highlights the memory allocation reductions the changes will achieve.

I don't have enough time to perfect the test for the time being, here is a simple case:

```pwsh
# windows
 git clone --bare https://github.com/git/git.git "$env:TEMP/git.git"
```

```go
func TestFindRef(t *testing.T) {
	now := time.Now()
	repo, err := git.PlainOpen("path/to/git")
	if err != nil {
		fmt.Fprintf(os.Stderr, "open error: %v\n", err)
		return
	}
	ref, err := repo.Reference("refs/heads/master", false)
	if err != nil {
		fmt.Fprintf(os.Stderr, "open error: %v\n", err)
		return
	}
	fmt.Fprintf(os.Stderr, "%s %s use time: %v\n", ref.Hash(), ref.Name(), time. Since(now))
}
```

refs:
```bash
# git for windows
git show-ref|wc -l
907
```

before:

![image](https://user-images.githubusercontent.com/6904176/234303672-bbc72440-18d8-4908-8fe4-809ce2417f6e.png)

>7580f92ffa970b9484ac214f7b53cec5e26ca4bc refs/heads/master use time: 1.6423ms

after:

![image](https://user-images.githubusercontent.com/6904176/234303954-9c664eb5-3c5b-4b4e-b23d-e20e237468af.png)

>7580f92ffa970b9484ac214f7b53cec5e26ca4bc refs/heads/master use time: 529.5µs