@pjbgf 2023-04-26
The benchmarks shows that the changes results in a more predictable time cost per op:

```
goos: linux
goarch: amd64
pkg: github.com/go-git/go-git/v5/storage/transactional
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz

/tmp/current
BenchmarkReference-16                 1        1807192975 ns/op        151665280 B/op    520597 allocs/op
BenchmarkReference-16                 1        2822611253 ns/op        148277904 B/op    482159 allocs/op
BenchmarkReference-16                 1        1871437889 ns/op        148332872 B/op    481993 allocs/op
BenchmarkReference-16                 1        2013824760 ns/op        147963112 B/op    482186 allocs/op
BenchmarkReference-16                 1        3188275241 ns/op        148189280 B/op    482150 allocs/op
BenchmarkReference-16                 1        1929030106 ns/op        148297832 B/op    481988 allocs/op

/tmp/after
BenchmarkReference-16                 1        1298677561 ns/op        150877264 B/op    520869 allocs/op
BenchmarkReference-16                 1        1170506662 ns/op        147997344 B/op    482935 allocs/op
BenchmarkReference-16                 1        1105562172 ns/op        147640968 B/op    482887 allocs/op
BenchmarkReference-16                 1        1056970154 ns/op        148353736 B/op    482919 allocs/op
BenchmarkReference-16                 1        1565437569 ns/op        148816872 B/op    482753 allocs/op
BenchmarkReference-16                 1        1095235494 ns/op        148314880 B/op    482941 allocs/op
```

benchstat
```
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
              │ /tmp/current │             /tmp/after             │
              │    sec/op    │   sec/op     vs base               │
BenchmarkReference-16    1.971 ± 62%   1.138 ± 38%  -42.27% (p=0.002 n=6)

              │ /tmp/current │          /tmp/after           │
              │     B/op     │     B/op      vs base         │
BenchmarkReference-16   141.4Mi ± 2%   141.5Mi ± 2%  ~ (p=0.818 n=6)

              │ /tmp/current │            /tmp/after             │
              │  allocs/op   │  allocs/op   vs base              │
BenchmarkReference-16    482.2k ± 8%   482.9k ± 8%  +0.16% (p=0.041 n=6)
```

Code used:
```go
func BenchmarkReference(b *testing.B) {
	repo, err := git.PlainClone(b.TempDir(), true, &git.CloneOptions{
		URL: "https://github.com/go-git/go-git.git",
	})
	if err != nil {
		b.Fatalf("open error: %v", err)
	}

	for i := 0; i < b.N; i++ {
		_, err := repo.Reference("refs/heads/master", false)
		if err != nil {
			b.Fatalf("open error: %v", err)
		}
	}
}
```